requirements.txt
flask
tensorflow-cpu
pillow
requests
numpy
python-multipart
